Strengthen modular product axioms
Factor mod_ax_mul into smaller helpers and add exact iff lemmas for
special residue cases. For prime moduli, encode product % p = 0 iff some
factor is 0 mod p. For parity, encode product % 2 = 1 iff all factors are
odd. Keep the generic modular congruence as a one-way implication for the
remaining cases.

Use unsigned modulus types and tighten const annotations in the modular
residue code. Add focused projection tests for prime-zero, parity,
powered-factor, composite-zero, and fallback congruence cases.

+++ src/qfn2l_cpp/lia_abstraction.cpp
@@ -814,8 +814,8 @@ smt::TermVec LiaAbstraction::mk_pow_axioms(const smt::Term& pure,
     if (_opts.modax > 2) {
         auto pure_val_opt = get_value(pure);
         if (pure_val_opt)
-            for (auto& ax : mod_ax_mul(_ctx, _opts.modax, {{root, exp, root_val}}, pure,
-                                       *pure_val_opt))
+            for (auto& ax : mod_ax_mul(_ctx, static_cast<unsigned int>(_opts.modax),
+                                       {{root, exp, root_val}}, pure, *pure_val_opt))
                 rv.push_back(ax);
     }
     return rv;
@@ -904,7 +904,7 @@ smt::TermVec LiaAbstraction::mk_mixed_mul_axioms(const smt::Term& pure,
         auto pv = get_value(pure);
         if (pv)
             for (auto& ax : mod_ax_mul(
-                     _ctx, _opts.modax,
+                     _ctx, static_cast<unsigned int>(_opts.modax),
                      {{root1, exp1, root1_val}, {root2, exp2, root2_val}}, pure, *pv))
                 rv.push_back(ax);
     }
